(cleanup) rename and move layers folder, clean up allLayers file and add comments, nothing broken
