+++ sentimentjs/baseWordsLayer/baseWordsList.js
