+++ sentimentjs/emoticonLayer/emojiRegex.js
