+++ sentimentjs/emoticonLayer/emoticonsList.js
